reduced int in favor of uint

+++ model/devices.go
@@ -11,13 +11,13 @@ type Device struct {
 	FirmwareVersion string `json:"firmwareVersion"`
 	Version         string `json:"version"`
 	HwVersion       string `json:"hwVersion"`
-	Status          int    `json:"status"`
-	StatusCategory  int    `json:"statusCategory"`
+	Status          uint64 `json:"status"`
+	StatusCategory  uint64 `json:"statusCategory"`
 	Site            string `json:"site"`
-	Compatible      int    `json:"compatible"`
+	Compatible      uint64 `json:"compatible"`
 	Sn              string `json:"sn"`
-	PortNum         int    `json:"portNum"`
-	LedSetting      int    `json:"ledSetting"`
+	PortNum         uint64 `json:"portNum"`
+	LedSetting      uint64 `json:"ledSetting"`
 	SnmpSeting      struct {
 		Location string `json:"location"`
 		Contact  string `json:"contact"`
@@ -31,37 +31,37 @@ type Device struct {
 	EchoServer      string `json:"echoServer"`
 	IP              string `json:"ip"`
 	Uptime          string `json:"uptime"`
-	UptimeLong      int    `json:"uptimeLong"`
-	CPUUtil         int    `json:"cpuUtil"`
-	MemUtil         int    `json:"memUtil"`
-	LastSeen        int64  `json:"lastSeen"`
+	UptimeLong      uint64 `json:"uptimeLong"`
+	CPUUtil         uint64 `json:"cpuUtil"`
+	MemUtil         uint64 `json:"memUtil"`
+	LastSeen        uint64 `json:"lastSeen"`
 	PortStats       []struct {
-		Port              int    `json:"port"`
+		Port              uint64 `json:"port"`
 		Name              string `json:"name"`
-		Type              int    `json:"type"`
-		Mode              int    `json:"mode"`
-		Status            int    `json:"status"`
-		Rx                int    `json:"rx"`
-		RxPkt             int    `json:"rxPkt"`
-		RxPktRate         int    `json:"rxPktRate"`
-		RxRate            int    `json:"rxRate"`
-		Tx                int    `json:"tx"`
-		TxPkt             int    `json:"txPkt"`
-		TxPktRate         int    `json:"txPktRate"`
-		TxRate            int    `json:"txRate"`
-		InternetState     int    `json:"internetState,omitempty"`
+		Type              uint64 `json:"type"`
+		Mode              uint64 `json:"mode"`
+		Status            uint64 `json:"status"`
+		Rx                uint64 `json:"rx"`
+		RxPkt             uint64 `json:"rxPkt"`
+		RxPktRate         uint64 `json:"rxPktRate"`
+		RxRate            uint64 `json:"rxRate"`
+		Tx                uint64 `json:"tx"`
+		TxPkt             uint64 `json:"txPkt"`
+		TxPktRate         uint64 `json:"txPktRate"`
+		TxRate            uint64 `json:"txRate"`
+		InternetState     uint64 `json:"internetState,omitempty"`
 		IP                string `json:"ip,omitempty"`
-		Speed             int    `json:"speed,omitempty"`
-		Duplex            int    `json:"duplex,omitempty"`
+		Speed             uint64 `json:"speed,omitempty"`
+		Duplex            uint64 `json:"duplex,omitempty"`
 		Proto             string `json:"proto,omitempty"`
-		WanIpv6Comptent   int    `json:"wanIpv6Comptent,omitempty"`
+		WanIpv6Comptent   uint64 `json:"wanIpv6Comptent,omitempty"`
 		WanPortIpv6Config struct {
-			Enable        int    `json:"enable"`
+			Enable        uint64 `json:"enable"`
 			Addr          string `json:"addr"`
 			Gateway       string `json:"gateway"`
 			PriDNS        string `json:"priDns"`
 			SndDNS        string `json:"sndDns"`
-			InternetState int    `json:"internetState"`
+			InternetState uint64 `json:"internetState"`
 		} `json:"wanPortIpv6Config,omitempty"`
 		WanPortIpv4Config struct {
 			IP       string `json:"ip"`
@@ -75,17 +75,17 @@ type Device struct {
 	} `json:"portStats"`
 	LanClientStats []struct {
 		LanName           string `json:"lanName"`
-		Vlan              int    `json:"vlan"`
+		Vlan              uint64 `json:"vlan"`
 		IP                string `json:"ip"`
-		Rx                int    `json:"rx"`
-		Tx                int    `json:"tx"`
-		ClientNum         int    `json:"clientNum"`
+		Rx                uint64 `json:"rx"`
+		Tx                uint64 `json:"tx"`
+		ClientNum         uint64 `json:"clientNum"`
 		LanPortIpv6Config struct {
 			Addr string `json:"addr"`
 		} `json:"lanPortIpv6Config"`
 	} `json:"lanClientStats"`
-	NeedUpgrade     bool  `json:"needUpgrade"`
-	Download        int64 `json:"download"`
-	Upload          int64 `json:"upload"`
-	NetworkComptent int   `json:"networkComptent"`
+	NeedUpgrade     bool   `json:"needUpgrade"`
+	Download        uint64 `json:"download"`
+	Upload          uint64 `json:"upload"`
+	NetworkComptent uint64 `json:"networkComptent"`
 }
